+++ pylibs/pylama/main.py
@@ -1,102 +1,39 @@
+""" Pylama shell integration.
+"""
+from __future__ import absolute_import, with_statement
+
 import fnmatch
+import logging
 import re
 import sys
+from argparse import ArgumentParser
 from os import getcwd, walk, path as op
 
-import logging
-from argparse import ArgumentParser
-
-from . import utils
+from . import utils, version
+from .core import DEFAULT_LINTERS, LOGGER, STREAM
 
 
-default_linters = 'pep8', 'pyflakes', 'mccabe'
-default_complexity = 10
-logger = logging.Logger('pylama')
-stream = logging.StreamHandler()
-logger.addHandler(stream)
-
-SKIP_PATTERN = '# nolint'
+DEFAULT_COMPLEXITY = 10
 
 
-def run(path, ignore=None, select=None, linters=default_linters, **meta):  # nolint
-    errors = []
-    ignore = ignore and list(ignore) or []
-    select = select and list(select) or []
+def shell(args=None, error=True):
+    """ Endpoint for console.
 
-    for lint in linters:
-        try:
-            linter = getattr(utils, lint)
-        except AttributeError:
-            logging.warning("Linter `{0}` not found.".format(lint))
-            continue
+    :return list: list of errors
+    :raise SystemExit:
 
-        try:
-            with open(path, "rU") as f:
-                code = f.read() + '\n\n'
-                params = parse_modeline(code)
-                params['skip'] = [False]
-                for line in code.split('\n'):
-                    params['skip'].append(line.endswith(SKIP_PATTERN))
+    """
+    curdir = getcwd()
+    if args is None:
+        args = sys.argv[1:]
 
-                if params.get('lint_ignore'):
-                    ignore += params.get('lint_ignore').split(',')
-
-                if params.get('lint_select'):
-                    select += params.get('lint_select').split(',')
-
-                if params.get('lint'):
-                    result = linter(path, code=code, **meta)
-                    for e in result:
-                        e['col'] = e.get('col') or 0
-                        e['lnum'] = e.get('lnum') or 0
-                        e['type'] = e.get('type') or 'E'
-                        e['text'] = "{0} [{1}]".format((e.get(
-                            'text') or '').strip()
-                            .replace("'", "\"").split('\n')[0], lint)
-                        e['filename'] = path or ''
-                        if not params['skip'][e['lnum']]:
-                            errors.append(e)
-
-        except IOError as e:
-            errors.append(dict(
-                lnum=0,
-                type='E',
-                col=0,
-                text=str(e)
-            ))
-        except SyntaxError as e:
-            errors.append(dict(
-                lnum=e.lineno or 0,
-                type='E',
-                col=e.offset or 0,
-                text=e.args[0]
-            ))
-            break
-
-        except Exception as e:
-            import traceback
-            logging.error(traceback.format_exc())
-
-    errors = [e for e in errors if _ignore_error(e, select, ignore)]
-    return sorted(errors, key=lambda x: x['lnum'])
-
-
-def _ignore_error(e, select, ignore):
-    for s in select:
-        if e['text'].startswith(s):
-            return True
-    for i in ignore:
-        if e['text'].startswith(i):
-            return False
-    return True
-
-
-def shell():
     parser = ArgumentParser(description="Code audit tool for python.")
-    parser.add_argument("path", nargs='?', default=getcwd(),
+    parser.add_argument("path", nargs='?', default=curdir,
                         help="Path on file or directory.")
     parser.add_argument(
         "--verbose", "-v", action='store_true', help="Verbose mode.")
+    parser.add_argument('--version', action='version',
+                        version='%(prog)s ' + version)
 
     split_csp_list = lambda s: list(set(i for i in s.split(',') if i))
 
@@ -108,9 +45,12 @@ def shell():
         type=split_csp_list,
         help="Select errors and warnings. (comma-separated)")
     parser.add_argument(
-        "--linters", "-l", default=','.join(default_linters),
+        "--linters", "-l", default=','.join(DEFAULT_LINTERS),
         type=split_csp_list,
-        help="Select linters. (comma-separated)")
+        help=(
+            "Select linters. (comma-separated). Choices are %s."
+            % ','.join(s for s in utils.__all__)
+        ))
     parser.add_argument(
         "--ignore", "-i", default='',
         type=split_csp_list,
@@ -120,84 +60,126 @@ def shell():
         type=lambda s: [re.compile(fnmatch.translate(p))
                         for p in s.split(',')],
         help="Skip files by masks (comma-separated, Ex. */messages.py)")
-    parser.add_argument("--complexity", "-c", default=default_complexity,
+    parser.add_argument("--complexity", "-c", default=DEFAULT_COMPLEXITY,
                         type=int, help="Set mccabe complexity.")
     parser.add_argument("--report", "-r", help="Filename for report.")
     parser.add_argument("--hook", action="store_true",
                         help="Install Git (Mercurial) hook.")
-    args = parser.parse_args()
+    parser.add_argument(
+        "--async", action="store_true",
+        help="Enable async mode. Usefull for checking a lot of files. "
+             "Dont supported with pylint.")
+    parser.add_argument(
+        "--options", "-o", default=op.join(curdir, 'pylama.ini'),
+        help="Select configuration file. By default is '<CURDIR>/pylama.ini'")
+    options = parser.parse_args(args)
+    actions = dict((a.dest, a) for a in parser._actions)
 
-    # Setup logger
-    logger.setLevel(logging.INFO if args.verbose else logging.WARN)
-    if args.report:
-        logger.removeHandler(stream)
-        logger.addHandler(logging.FileHandler(args.report, mode='w'))
+    # Read options from configuration file
+    from .inirama import Namespace
 
-    if args.hook:
+    config = Namespace()
+    config.default_section = 'main'
+    config.read(options.options)
+    for k, v in config.default.items():
+        action = actions.get(k)
+        if action:
+            LOGGER.info('Find option %s (%s)', k, v)
+            name, value = action.dest, action.type(v)\
+                if callable(action.type) else v
+            if action.const:
+                value = bool(int(value))
+            setattr(options, name, value)
+
+    # Setup LOGGER
+    LOGGER.setLevel(logging.INFO if options.verbose else logging.WARN)
+    if options.report:
+        LOGGER.removeHandler(STREAM)
+        LOGGER.addHandler(logging.FileHandler(options.report, mode='w'))
+    LOGGER.info('Try to read configuration from: ' + options.options)
+
+    # Install VSC hook
+    if options.hook:
         from .hook import install_hook
-        return install_hook(args.path)
+        install_hook(options.path)
 
-    paths = [args.path]
+    else:
 
-    if op.isdir(args.path):
-        paths = []
-        for root, _, files in walk(args.path):
-            paths += [op.join(root, f) for f in files if f.endswith('.py')]
+        paths = [options.path]
 
-    check_files(
-        paths,
-        rootpath=args.path,
-        skip=args.skip,
-        frmt=args.format,
-        ignore=args.ignore,
-        select=args.select,
-        linters=args.linters,
-        complexity=args.complexity,
-    )
+        if op.isdir(options.path):
+            paths = []
+            for root, _, files in walk(options.path):
+                paths += [
+                    op.relpath(op.join(root, f), curdir)
+                    for f in files if f.endswith('.py')]
+
+        return check_files(
+            paths,
+            async=options.async,
+            rootpath=curdir,
+            skip=options.skip,
+            frmt=options.format,
+            ignore=options.ignore,
+            select=options.select,
+            linters=options.linters,
+            complexity=options.complexity,
+            config=config,
+            error=error,
+        )
 
 
-def check_files(paths, rootpath=None, skip=None, frmt="pep8",
-                select=None, ignore=None, linters=default_linters,
-                complexity=default_complexity):
+def check_files(paths, rootpath=None, skip=None, frmt="pep8", async=False,
+                select=None, ignore=None, linters=DEFAULT_LINTERS,
+                complexity=DEFAULT_COMPLEXITY, config=None, error=True):
+    """ Check files.
+
+    :return list: list of errors
+    :raise SystemExit:
+
+    """
+    from .tasks import async_check_files
+
     rootpath = rootpath or getcwd()
     pattern = "%(rel)s:%(lnum)s:%(col)s: %(text)s"
     if frmt == 'pylint':
         pattern = "%(rel)s:%(lnum)s: [%(type)s] %(text)s"
 
-    errors = []
-    for path in skip_paths(skip, paths):
-        logger.info("Parse file: %s" % path)
-        errors = run(path, ignore=ignore, select=select,
-                     linters=linters, complexity=complexity)
-        for error in errors:
-            try:
-                error['rel'] = op.relpath(
-                    error['filename'], op.dirname(rootpath))
-                error['col'] = error.get('col', 1)
-                logger.warning(pattern, error)
-            except KeyError:
-                continue
+    params = dict()
+    if config:
+        for key, section in config.sections.items():
+            if key != config.default_section:
+                mask = re.compile(fnmatch.translate(key))
+                params[mask] = prepare_params(section)
 
-    sys.exit(int(bool(errors)))
-
-
-MODERE = re.compile(
-    r'^\s*#\s+(?:pymode\:)?((?:lint[\w_]*=[^:\n\s]+:?)+)', re.I | re.M)
-
-
-def skip_paths(skip, paths):
+    work_paths = []
     for path in paths:
         if skip and any(pattern.match(path) for pattern in skip):
+            LOGGER.info('Skip path: %s', path)
             continue
-        yield path
+        work_paths.append(path)
+
+    errors = async_check_files(
+        work_paths, async=async, rootpath=rootpath, ignore=ignore,
+        select=select, linters=linters, complexity=complexity, params=params)
+
+    for er in errors:
+        LOGGER.warning(pattern, er)
+
+    if error:
+        sys.exit(int(bool(errors)))
+
+    return errors
 
 
-def parse_modeline(code):
-    seek = MODERE.search(code)
-    params = dict(lint=1)
-    if seek:
-        params = dict(v.split('=') for v in seek.group(1).split(':'))
-        params['lint'] = int(params.get('lint', 1))
+def prepare_params(section):
+    """ Parse modeline params from configuration.
+
+    :return dict: Linter params.
+
+    """
+    params = dict(section)
+    params['lint'] = int(params.get('lint', 1))
     return params