+++ pylibs/pymode/lint.py
@@ -1,8 +1,17 @@
+""" Pylama support. """
+from __future__ import absolute_import
+
+import fnmatch
+import json
 import locale
+from os import path as op
+from re import compile as re
 
-from pylama.main import run
+from pylama.core import run
+from pylama.inirama import Namespace
+from pylama.main import prepare_params
 
-from .interface import get_option, get_var, get_current_buffer, command
+from . import interface
 from .queue import add_task
 
 
@@ -13,51 +22,73 @@ except AttributeError:
 
 
 def check_file():
-    checkers = get_option('lint_checker').split(',')
+    """ Check current buffer. """
+    checkers = interface.get_option('lint_checker').split(',')
+    buf = interface.get_current_buffer()
+
+    # Check configuration from `pymode.ini`
+    curdir = interface.eval_code('getcwd()')
+    config = Namespace()
+    config.default_section = 'main'
+    config.read(op.join(curdir, 'pylama.ini'), op.join(curdir, 'pymode.ini'))
 
     ignore = set([
         i for i in (
-            get_option('lint_ignore').split(',') +
-            get_var('lint_ignore').split(','))
-        if i
-    ])
-    select = set([
-        s for s in (
-            get_option('lint_select').split(',') +
-            get_var('lint_select').split(','))
-        if s
+            interface.get_option('lint_ignore').split(',') +
+            interface.get_var('lint_ignore').split(',') +
+            config.default.get('ignore', '').split(',')
+        ) if i
     ])
 
-    buf = get_current_buffer()
-    complexity = int(get_option('lint_mccabe_complexity') or 0)
+    select = set([
+        s for s in (
+            interface.get_option('lint_select').split(',') +
+            interface.get_var('lint_select').split(',') +
+            config.default.get('select', '').split(',')
+        ) if s
+    ])
+
+    complexity = int(interface.get_option('lint_mccabe_complexity') or 0)
+
+    params = dict()
+    relpath = op.relpath(buf.name, curdir)
+    for mask in config.sections:
+        mask_re = re(fnmatch.translate(mask))
+        if mask_re.match(relpath):
+            params.update(prepare_params(config[mask]))
+
+    if relpath in config:
+        params = prepare_params(config[relpath])
 
     add_task(
         run_checkers,
         callback=parse_result,
         title='Code checking',
 
-        checkers=checkers,
-        ignore=ignore,
-        buf=buf,
-        select=select,
-        complexity=complexity)
+        # params
+        checkers=checkers, ignore=ignore, buf=buf, select=select,
+        complexity=complexity, config=params,
+    )
 
 
 def run_checkers(checkers=None, ignore=None, buf=None, select=None,
-                 complexity=None, callback=None):
+                 complexity=None, callback=None, config=None):
+    """ Run pylama code.
 
-    filename = buf.name
-    result = []
+    :return list: errors
 
-    pylint_options = '--rcfile={0} -r n'.format(get_var('lint_config')).split()
+    """
+    pylint_options = '--rcfile={0} -r n'.format(
+        interface.get_var('lint_config')).split()
 
-    return run(filename, ignore=ignore, select=select, linters=checkers,
-                 pylint=pylint_options, complexity=complexity)
+    return run(
+        buf.name, ignore=ignore, select=select, linters=checkers,
+        pylint=pylint_options, complexity=complexity, config=config)
 
 
 def parse_result(result, buf=None, **kwargs):
-    command(('let g:qf_list = {0}'.format(repr(result)).replace(
-        '\': u', '\': ')))
-    command('call pymode#lint#Parse({0})'.format(buf.number))
+    """ Parse results. """
+    interface.command('let g:qf_list = ' + json.dumps(result))
+    interface.command('call pymode#lint#Parse({0})'.format(buf.number))
 
 # pymode:lint_ignore=W0622